reorganizando os componentes do menu
