--- src/components/snacks.js
@@ -1,7 +1,0 @@
-const Snacks = () => {
-    return (
-     <p> Grilled Cheese Sandwich $10</p>
-    )
-}
-
-export default Snacks;
